changement final

+++ src/main/java/com/abirch/security/user/PermissionRepository.java
@@ -2,6 +2,11 @@ package com.abirch.security.user;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PermissionRepository extends JpaRepository<Permission, Integer> {
+import java.util.List;
+import java.util.Set;
+
+public interface PermissionRepository extends JpaRepository<Permission, Long> {
+  //  List<Permission> findAllById(Set<Long> ids);
+
     // Tu peux ajouter des méthodes personnalisées ici si nécessaire
 }
